Refactor some code into modules; Move `11px` to *popup.css*; Increase search field border radius; Change title `Doki Theme for Firefox` to `Doki Theme`; Updates browser action logo to change color based on currently selected theme; Update CHANGELOG.md; Update README.md; Removes unused namespace from svg;
