+++ firefoxThemes/popup/index.html
@@ -6,7 +6,7 @@
   <link rel="stylesheet" href="popup.css"/>
   <meta content="width=device-width, initial-scale=1.0" name="viewport">
   <!--Popup Script-->
-  <script defer src="popup.js"></script>
+  <script type="module" src="popup.js" defer ></script>
 </head>
 <body class="popup-body">
 <header class="popup-header">
