+++ firefoxThemes/popup/popup.css
@@ -9,7 +9,7 @@
   --button-font-color: #47494b;
   --base-background-color: #2f363c;
   --switch-shadow-color: #777;
-  --doki-shadow: '11px #fff';
+  --doki-shadow: #fff;
 }
 
 /*Classes*/
@@ -99,7 +99,7 @@ img {
 
 img:hover {
   cursor: pointer;
-  filter: drop-shadow(0px 0px var(--doki-shadow));
+  filter: drop-shadow(0px 0px 11px var(--doki-shadow));
 }
 
 /*Waifu Option Styles*/
