+++ firefoxThemes/waifus/index.html
@@ -8,7 +8,7 @@
   <!--Styles for new tab page-->
     <link href="tab.css" rel="stylesheet">
     <!--Search Capabilities-->
-    <script src="tab.js" defer></script>
+    <script type="module" src="tab.js" defer></script>
     <!--Adds/Removes Waifu-->
     <script src="waifu.js" defer></script>
 </head>
